pyproject.toml
[build-system]
requires = ["maturin>=1.11,<2.0"]
build-backend = "maturin"

[project]
name = "python-matchit"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]

[dependency-groups]
dev = [
    "maturin>=1.11.5",
    "python-matchit",
]

[tool.uv.workspace]
members = [
    ".",
]

[tool.uv.sources]
python-matchit = { workspace = true }
